backend_v2/pyproject.toml
[tool.poetry]
name = "trackrat-v2"
version = "2.0.0"
description = "TrackRat V2 Backend - Simplified train tracking system"
authors = ["TrackRat Team"]
readme = "README.md"
packages = [{include = "trackrat", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
aiosqlite = "^0.19.0"

# HTTP Client
httpx = {extras = ["http2"], version = "^0.27.0"}

# Task Scheduling
apscheduler = "^3.10.4"

# Utilities
python-dateutil = "^2.8.2"
pytz = "^2024.1"
structlog = "^24.1.0"

# Monitoring
prometheus-client = "^0.19.0"

# Apple Push Notifications
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"

# Type Checking
mypy = "^1.8.0"
types-python-dateutil = "^2.8.19"
types-pytz = "^2024.1"
sqlalchemy = {extras = ["mypy"], version = "^2.0.25"}

# Code Quality
black = "^24.1.1"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = "apscheduler.*"
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/trackrat"]
omit = ["*/tests/*", "*/migrations/*"]